openxr: Migrate replay state to individual files
Session and Swapchain get their own source and headers.
